Moved test helpers back to `src/`
